Minor change. Cast the number argument to int in call to CLSSIS3820Scaler::measureDarkCurrent().

+++ source/ui/CLS/CLSSIS3820ScalerDarkCurrentWidget.cpp
@@ -97,6 +97,6 @@ void CLSSIS3820ScalerDarkCurrentWidget::onCollectButtonClicked()
 	double secondsEntered = timeEntry_->value() / MILLISECONDS_PER_SECOND;
 
 	if (scaler_ && secondsEntered > 0) {
-		scaler_->measureDarkCurrent(secondsEntered);
+		scaler_->measureDarkCurrent((int)secondsEntered);
 	}
 }
